Time: 5 ms (75.27%), Space: 43.6 MB (83.18%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.java
@@ -2,8 +2,8 @@ class Solution
 {
     public boolean isPalindrome(int x) 
     {
-        long rev = 0;
-        long temp = x;
+        int rev = 0;
+        int temp = x;
         
         if (x < 0) 
         {
@@ -12,7 +12,7 @@ class Solution
 
         while (temp != 0) 
         {
-            int d = (int)(temp % 10);
+            int d = (temp % 10);
             rev = rev * 10 + d;
             temp = temp/10;
         }
